fix: Add nats token to function output collector
